requirements.txt
config==0.5.1
fastapi==0.110.1
fastapi_health==0.4.0
middleware==1.2.3
python-decouple==3.8
pytube==15.0.0
starlette==0.37.2
starlette_prometheus==0.9.0
uvicorn==0.29.0